requirements.txt
flask
flask_sqlalchemy
requests
python-dotenv
pytest
flask-migrate